UBERF-5586: load saved messages using one query and save in store
