+++ plugins/activity-resources/src/components/Activity.svelte
@@ -17,10 +17,12 @@
   import { Doc, Ref, SortingOrder } from '@hcengineering/core'
   import { createQuery, getClient } from '@hcengineering/presentation'
   import { Component, Grid, Label, Lazy, Spinner } from '@hcengineering/ui'
-  import ActivityExtensionComponent from './ActivityExtension.svelte'
+  import { onMount } from 'svelte'
 
+  import ActivityExtensionComponent from './ActivityExtension.svelte'
   import ActivityFilter from './ActivityFilter.svelte'
   import { combineActivityMessages } from '../activityMessagesUtils'
+  import { loadSavedMessages } from '../activity'
 
   export let object: Doc
   export let showCommenInput: boolean = true
@@ -67,6 +69,10 @@
   }
 
   $: void updateActivityMessages(object._id, isNewestFirst ? SortingOrder.Descending : SortingOrder.Ascending)
+
+  onMount(() => {
+    loadSavedMessages()
+  })
 </script>
 
 <div class="antiSection-header high mt-9" class:invisible={transparent}>
