+++ plugins/chunter-resources/src/components/chat-message/ChatMessagePopup.svelte
@@ -14,13 +14,15 @@
 // limitations under the License.
 -->
 <script lang="ts">
-  import { createEventDispatcher } from 'svelte'
+  import { createEventDispatcher, onMount } from 'svelte'
   import { Doc, Ref, SortingOrder } from '@hcengineering/core'
   import { createQuery } from '@hcengineering/presentation'
   import activity from '@hcengineering/activity'
   import chunter, { ChatMessage } from '@hcengineering/chunter'
   import { closeTooltip, Label, Lazy, Spinner, resizeObserver, MiniToggle } from '@hcengineering/ui'
   import { ObjectPresenter, DocNavLink } from '@hcengineering/view-resources'
+  import { loadSavedMessages } from '@hcengineering/activity-resources'
+
   import ChatMessageInput from './ChatMessageInput.svelte'
   import ChatMessagePresenter from './ChatMessagePresenter.svelte'
 
@@ -51,6 +53,10 @@
   $: if (isTextMode) {
     dispatch('tooltip', { kind: 'popup' })
   }
+
+  onMount(() => {
+    loadSavedMessages()
+  })
 </script>
 
 <div class="commentPopup-container">
