+++ plugins/notification-resources/src/components/inbox/Inbox.svelte
@@ -37,7 +37,8 @@
   import { Ref, WithLookup } from '@hcengineering/core'
   import { ViewletSelector } from '@hcengineering/view-resources'
   import activity, { ActivityMessage } from '@hcengineering/activity'
-  import { isReactionMessage } from '@hcengineering/activity-resources'
+  import { isReactionMessage, loadSavedMessages } from '@hcengineering/activity-resources'
+  import { onMount } from 'svelte'
 
   import { inboxMessagesStore, InboxNotificationsClientImpl } from '../../inboxNotificationsClient'
   import Filter from '../Filter.svelte'
@@ -250,6 +251,10 @@
     { minSize: 30, maxSize: 50, size: 40, float: 'navigator' },
     { size: 'auto', minSize: 30, maxSize: 'auto', float: undefined }
   ])
+
+  onMount(() => {
+    loadSavedMessages()
+  })
 </script>
 
 <ActionContext
